+++ extensions/shared-youtube/library/src/main/java/app/morphe/extension/shared/returnyoutubedislike/requests/RYDVoteData.java
@@ -1,10 +1,3 @@
-/*
- * Copyright 2026 Morphe.
- * https://github.com/MorpheApp/morphe-patches
- *
- * See the included NOTICE file for GPLv3 Section 7 terms that apply to this code.
- */
-
 package app.morphe.extension.shared.returnyoutubedislike.requests;
 
 import static app.morphe.extension.shared.returnyoutubedislike.ReturnYouTubeDislike.Vote;
